HBX-1590: Move class 'org.hibernate.cfg.JDBCReaderFactory' to 'org.hibernate.tool.internal.reveng.JdbcReaderFactory'
