+++ main/src/main/java/org/hibernate/tool/internal/reveng/JdbcReaderFactory.java
@@ -1,7 +1,8 @@
-package org.hibernate.cfg;
+package org.hibernate.tool.internal.reveng;
 
 import java.util.Properties;
 
+import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.reveng.JDBCReader;
 import org.hibernate.engine.jdbc.connections.spi.ConnectionProvider;
 import org.hibernate.engine.jdbc.spi.JdbcServices;
@@ -11,7 +12,7 @@ import org.hibernate.tool.api.dialect.MetaDataDialect;
 import org.hibernate.tool.api.dialect.MetaDataDialectFactory;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
 
-final public class JDBCReaderFactory {
+final public class JdbcReaderFactory {
 
 	
 	public static JDBCReader newJDBCReader(
